fix: add /var/www to git safe directory
While tweaking the setup and Docker permissions, the show branch feature
stopped working. It seems that the /var/www directory which is the mounted
directory from the host machine, was no longer recognized as a safe directory
by git. Let's add it to the safe directory list when running the Docker file.