fix: make public directories writable by www-data and hostgroup
This was in the original docker_run.sh script, but removed thinking it was
no longer needed.

This adds it back with some tweaks. Instead of changing the permissions
of all directories and files, we'll only change the permissions of the
directories. Otherwise, it seems like git shows all the files as changed
thanks to the permission changes.