feat(docker): extract development and discourse services to separate files
- Introduced docker-compose.dev.yml for development services including
phpMyAdmin and Mailhog.
- Added docker-compose.discourse.yml for Discourse services including
PostgreSQL, Redis, and Sidekiq.
- Removed existing phpMyAdmin and Discourse configurations from
docker-compose.yml to streamline service management.

By extracting these services into separate files we can build up the
stack on a case by case basis. This will free up host machines resources
and make it easier to manage the development environment.

I will be creating some development scripts in the future to help
with the setup and tearing down of the development environment.