refactor(docker): update comments in docker-compose files
This cleans up the comments in the docker-compose files since I had
originally created each docker-compose file by `cp docker-compose.yml`.

Some of the comments were removed in some of the files because they
referenced services that were not present in the file. Additionally,
I refactored some of the comments to be more linear and easier to read.