feat(task): add Taskfile for Docker container management
Taskfile.yml was chosen over Makefile primarily because its YAML syntax
is significantly easier to read and understand. This improves
maintainability and makes it more approachable to folks who may not
be familiar with the syntax of Makefiles. Additionally, Taskfile is
also designed as a general task runner, unlike the build-focused
nature of Makefiles. This makes it a better fit for managing
diverse workflows, automating steps, and helper functions
beyond just compilation. For now though, we will only be using it
to manage the Docker containers while we work on refactoring the app
setup process.

That said, we could have gone with bash scripts to manage these tasks,
but bash scripts become increasingly difficult to manage and read
as the logic grows. Writing intricate functions, handling argument
parsing, implementing error checking, and managing dependencies between
steps often leads to bash code that is verbose, hard to follow, and
challenging to debug and maintain. Taskfile provides a more structured
and familiar syntax to orchestrate the tasks.