feat(task): add tasks for managing all docker containers
Taskfiles allow the use of wildcards to pass arguments as part of the
task name. This allows us to reduce the number of tasks since we don't
need to create a task for each combination of containers we want to start
or stop. For example, docker:up:dev, docker:up:discourse, etc.

One caveat is that Taskfiles do not currently support lazy variable
lookup. As such, we can't save the output of a dependent task to a
variable and reuse it in a subsequent task. This is a bit of a pain,
but for our use case, we can live with it. The workaround then was to
use a map variable to act as a switch statement to determine which stack
to start or stop in line. This requires the map experimental feature to be
enabled, which we have done via the .env file.