refactor: make Task a first-class citizen
This updates the docs to exclusively reference Task commands and removes the
alternative Docker commands. As such, we are indicating that Task is
now mandatory.

Following this, I've moved the Task experimental env variable to a dedicated
`.taskrc.yml` file. This will be more appropriate and easier to manage
then having it in the `.env.example` file.