refactor(docker): remove root user declaration
By default, Docker uses the root user as such
the need to specify the user in the docker-compose.yml file
and the usage of sudo in the Dockerfile are unnecessary.

We will be adding a non-root user in a later commit
to handle the file permissions of the mounted
volumes.