docs(local-development): update environment configuration instructions
Clarified the automatic environment setup process when using Task commands.
As such we no longer need to manually copy the `.env` file and instead use the
`task env:generate` command to generate the `.env` file.

There are some additional notes about these changes and how its meant to
be used in the context of Docker.