fix(composer): resolve composer deps from laravel 10x upgrade
There were three main packages that needed to be manually fixed:
- wouternl/laravel-drip
- mcamara/laravel-localization
- addwiki/mediawiki-api-base

Laravel Drip doesn't seem to be used according to composer-unused. In
addition, we do not need to have marketing emails for the main app. As
such, we will remove the package and everything related to it. This
includes needing to create a migration to remove it from the DB.

Laravel Localization and Mediawiki API Base were also manually bumbed.
However, some issues occurred along the way.

The patch files we have in the repo could not be applied and may also
no longer be needed. I removed them from the composer.patches.json file,
but kept the actual patch files in the patches/ directory in case we need
to refer to them later.

Lastly, the Mediawiki API Base package depends on psr/log v3 but
php-coveralls requires psr/log v1 or v2. There is an open issue to bump
the psr/log version to v3 for php-coveralls, but its been close to a
year since the issue was opened, with no activity on it. I decided to
drop php-coveralls for that reason as well as it is not necessary for
our forked application.