fix(package.json): add name field
This was an issue prior to running Laravel Shift. There is no name field
in the package.json file so when we generate the package-lock.json file
it would use the current directory name for the package name. This becomes
an issue when running Docker as the working directory is set to /var/www
which alters the package-lock.json file.

Let's explicitly set the name field in the package.json file to "restarters"
so that the package-lock.json file is generated with the correct package name
regardless of the working directory.