fix(js-translations): readd translation builder to vite workflow
We were previously execing the artisan laravel-js-translations command
in the build step for Laravel Mix, but Laravel Shift did not carry this
over to the Vite workflow. Let's readd it.

Instead of copying it 1:1 from the Laravel Mix workflow, we will reimplement
the intent of the command in a way that works with Vite for production.

We will add a `prebuild` script and `postbuild` script to the package.json
that will mimic the behavior of what we had in the Laravel Mix workflow.

However, for development, we will use vite-plugin-watch-and-run to watch
the lang directory and run the translation builder when it changes. This
solves a prior problem where changes to `lang` were not picked up by
Webpack and therefore needed to restart the node process. Now, Vite
itself will watch the lang directory and trigger a rebuild when it
changes.

One thing to note is there is a conflict between packages since there is
a mix of CJS and ESM. Vite treats packages as ESM by default, but the
laravel-vite-plugin is CJS. All packages are moving towards ESM so we will
set the type of the project to "module" in package.json and will need
to use laravel.default() instead of laravel() in vite.config.js to ensure
it works.