chore: clean up last bits of Mix migration
This removes the last remaining Mix files from the codebase.

We won't be using BrowserSync, if on the off chance we need it in the
future we will add it back and implement it properly for Vite.

Additionally, Vite creates a new `hot` file with its dev server so we
will need to add it to the .gitignore file while removing the old
mix-manifest.json file entry.