breaking(composer): get package conflicts resolved to run composer update
Laravel Shift does not automatically run composer update, so we need to
run it manually. That said, there were some errors with the current state
of the composer.json file that needed to be resolved first.
- laravelcollective/html was replaced with spatie/laravel-html since its
been deprecated and is no longer maintained.
    - There are not a lot of uses for the original package, so we may
    want to consider replacing it with raw HTML.
- intervention/image was updated to version 3.11.2 to resolve a conflict
with laravel/framework at version 11.44.0.
- osteel/openapi-httpfoundation-testing was updated to version 0.11.0 to
resolve a conflict with laravel/framework at version 11.44.0.

Although we were able to get composer update to run, there are still
some issues that need to be resolved since we've updated major versions
of some packages and replaced a package with another.