fix(login): manually declare User model path
Laravel Shift 11 was a bit more aggressive with the changes since the
directory structure is now different. As such, there are certain expectations
that are not met. One of them is that it expects models to be in the `app/Models`
directory.

We will address all models in the future, but for now let's explicity
declare the User model path.