fix(mediawiki): update Mediawiki API to use new methods
A lot of the Mediawiki API methods have been changed in version 3.
https://addwiki.github.io/v3/mediawiki-api-base/quickstart.html. As such,
we need to update the usage in our codebase.

That said, one caveat to this is the cookie handling. We removed
the patching of the private `getClient()` method in
https://github.com/iFixit/restarters/blob/c6338441ff6ce248246227ebe15cfe75e891d784/patches/composer.patches.json
The patch would convert getClient() to a public method and allow us
access to the cookies.

To resolve this, we will use the Reflection class to access the private
property. This is better than having a patch that may break on updates.