refactor(routes): complete transition to AppServiceProvider
Laravel 11 introduced a new way to handle routes and providers. The
Service providers were consolidated with Laravel Shift and as such a lot
of the code in RouteServiceProvider was moved to AppServiceProvider.

There were a few other changes that were needed to ensure the same
functionality was maintained. As such, we can remove RouteServiceProvider
entirely.