fix(dependencies): get compatible Laravel 12 dependencies
Laravel 12 does not run composer update or npm install so we need to
do it manually.

That said, there is version conflict with the laravel-discourse-sso package.
It is currently not compatible with Laravel 12 and the pull to fix it
has not been merged or touched since February. There is however a fork
by Laravel Shift that has the fix.

As such, we will use that fork in the composer.json file.

Reference:
https://github.com/spinen/laravel-discourse-sso/pull/52