feat(taskfile): add task to initialize the test database
We will need to run this task before running the tests, but we would be
able to utilize the same MySQL container for tests.

We will pull in the .env.testing file to override the .env.docker vars
that point to the dev db. That said, there seems to be a current bug
with Task where task specific dotenvs values are not accessible with the
Go template engine, while the global dotenv values are. As such,
we need to use the $VAR_NAME syntax to access the task specific dotenv
values.