feat(tests): add helper function for API test authentication
This adds a helper function to create a user with a unique API token.
This is similar to what we did in GroupCreateTest but we would need to
do this for all tests that used a static and non-unique token. This
all stems from switching to a transaction-based test setup and tests
making API calls which bypass the transaction event system -- can't
undo DB transactions if queries are made outside the testing context.

As such, this is a helper function that can be used in all tests that
need to create a specific user role for API tests and ensures that
the token won't cause conflicts.