revert(20d5726): Laravel Mix to Vite
I couldn't get Vite to work properly with the current dependencies.
A lot of the packages are out of date and as such hard to use as
ESM modules.

I tried patching, aliasing, altering packages, but it was one issue
after another. The only solution I have not tried to is fully update
the entire Vue codebase and packages, but that is a lot of work to do
at the moment.

For now, we will revert back to using Laravel Mix.