fix(docker): use watch-poll instead
There are some issues with the watch command on different operating systems.
The watch command relies on the File System Events API to detect changes in
the file system but in Windows this may be bypassed because of the multiple
layers of virtualization. Therefore a polling mechanism is more reliable.